added tests and new structure
