+++ src/dailyReportMethods/extraction/dataExtractionUtilities.js
